backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
Pillow==10.0.0
psycopg2-binary==2.9.6
gunicorn==21.1.0
python-dotenv==1.0.0